move decodeFromStorage, decodeFromStack, decodeFromMemory to ValueType
