+++ src/solidity/types/FixedByteArray.js
@@ -1,5 +1,4 @@
 'use strict'
-var util = require('./util')
 var ValueType = require('./ValueType')
 
 class FixedByteArray extends ValueType {
@@ -7,23 +6,8 @@ class FixedByteArray extends ValueType {
     super(1, storageBytes, 'bytesX')
   }
 
-  decodeFromStorage (location, storageContent) {
-    var value = util.extractHexValue(location, storageContent, this.storageBytes)
-    return '0x' + value.toUpperCase()
-  }
-
-  decodeFromStack (stackDepth, stack, memory) {
-    if (stack.length - 1 < stackDepth) {
-      return '0x'
-    } else {
-      var value = stack[stack.length - 1 - stackDepth]
-      return '0x' + value.substr(2, 2 * this.storageBytes).toUpperCase()
-    }
-  }
-
-  decodeFromMemory (offset, memory) {
-    var value = memory.substr(offset, 64)
-    return util.extractHexByteSlice(value, this.storageBytes, 0).toUpperCase()
+  decodeValue (value) {
+    return '0x' + value.substr(0, 2 * this.storageBytes).toUpperCase()
   }
 }
